19-remove-nth-node-from-end-of-list/solution.py
"""19. Remove Nth Node From End of List.

Given the head of a linked list, remove the nth node from the end of the list
and return its head.

Constraints:
    The number of nodes in the list is sz.
    1 <= sz <= 30
    0 <= Node.val <= 100
    1 <= n <= sz
"""

from typing import Optional


class ListNode:
    def __init__(self, val=0, next: Optional["ListNode"] = None):
        self.val = val
        self.next = next


class Solution:
    def removeNthFromEnd(
        self,
        head: Optional[ListNode],
        n: int,
    ) -> Optional[ListNode]:
        slow, fast = head, head

        while n and fast:
            fast = fast.next
            n -= 1

        if not fast:
            return head.next

        while fast.next:
            slow = slow.next
            fast = fast.next

        slow.next = slow.next.next
        return head
